logout functionality

+++ Resources/item.py
@@ -2,7 +2,7 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from schemas import ItemSchema, ItemUpdateSchema
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -24,6 +24,12 @@ class Item(MethodView):
 
     @jwt_required()
     def delete(self, item_id):
+
+        #using jwt claims
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required")
+
         item = ItemModel.query.get_or_404(item_id)
 
         db.session.delete(item)
